refactor: move cosntant adn function
