+++ openclaw-platform/.env.example
@@ -1,5 +1,6 @@
 # Mistral (primary receipt parser)
 MISTRAL_API_KEY=your_mistral_api_key
+MISTRAL_API_BASE=https://api.mistral.ai
 RECEIPT_MODEL=mistral-small-latest
 
 # Optional OpenAI fallback (unused by default parser)
@@ -7,6 +8,7 @@ OPENAI_API_KEY=
 
 # Telegram (used by OpenClaw channel runtime)
 TELEGRAM_BOT_TOKEN=your_telegram_bot_token
+TELEGRAM_API_BASE=https://api.telegram.org
 
 # Google Sheets
 GOOGLE_APPLICATION_CREDENTIALS=/absolute/path/to/service-account.json
@@ -18,9 +20,15 @@ RECEIPT_SHEET_MONTHLY=monthly_breakdown
 RECEIPT_MAX_PDF_PAGES=3
 RECEIPT_ACCEPT_PDF=false
 RECEIPT_STRICT_MEMORY_ONLY=false
+RECEIPT_JOURNAL_PATH=/absolute/path/to/chief-of-staff/obsidian-vault/md-db/receipt-journal.md
+RECEIPT_CONFIRMATION_TTL_MS=1800000
 NODE_ENV=development
 TZ=Asia/Jakarta
 OPENCLAW_HOME=/absolute/path/to/chief-of-staff/openclaw-platform/.openclaw-home
 
 OPENROUTER_API_KEY=xxxxxx
 GROQ_API_KEY=xxxxx
+
+MISTRAL_API_BASE=https://api.mistral.ai
+TELEGRAM_API_BASE=https://api.telegram.org
+RECEIPT_JOURNAL_PATH=/Users/nurrizky/dev/chief-of-staff/obsidian-vault/md-db/receipt-journal.md
