+++ openclaw-platform/src/config/env.ts
@@ -14,8 +14,10 @@ const BoolLikeSchema = z.preprocess((value) => {
 const EnvSchema = z.object({
   OPENAI_API_KEY: z.string().min(1).optional(),
   MISTRAL_API_KEY: z.string().min(1),
+  MISTRAL_API_BASE: z.string().url().default("https://api.mistral.ai"),
   RECEIPT_MODEL: z.string().default("mistral-small-latest"),
   TELEGRAM_BOT_TOKEN: z.string().min(1).optional(),
+  TELEGRAM_API_BASE: z.string().url().default("https://api.telegram.org"),
   GOOGLE_APPLICATION_CREDENTIALS: z.string().min(1),
   RECEIPT_SPREADSHEET_ID: z.string().min(1),
   RECEIPT_SHEET_RAW: z.string().default("receipts_raw"),
@@ -23,6 +25,8 @@ const EnvSchema = z.object({
   RECEIPT_MAX_PDF_PAGES: z.coerce.number().int().min(1).max(10).default(3),
   RECEIPT_ACCEPT_PDF: BoolLikeSchema.default(false),
   RECEIPT_STRICT_MEMORY_ONLY: BoolLikeSchema.default(false),
+  RECEIPT_JOURNAL_PATH: z.string().min(1),
+  RECEIPT_CONFIRMATION_TTL_MS: z.coerce.number().int().min(1).default(30 * 60 * 1000),
   NODE_ENV: z.string().default("development"),
   OPENCLAW_HOME: z.string().optional(),
   TZ: z.string().default("Asia/Jakarta")
